fix: use `esbuild.transformSync` (#6)
The `buildSync` export is more suitable for bundling, while `transformSync` is meant for single-file transformations.

A nice side effect of using `transformSync` is that the `sources` array in the source map no longer contains relative paths. But we still nullify the `sourcesContent` to reduce cache size (when Jest caches the source map).